gemini-engineering-doc/requirements.txt
streamlit
google-adk
python-dotenv
google-cloud-aiplatform
google-generativeai